+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Changes
 
 * Update to Go 1.16
+* Add mmlsfileset collector
 
 ## 1.2.0 / 2021-04-15